Add Closest Airport to weather and fix styling
In the lab on Tuesday the class discussed styling changes that would
improve the look of the weather tab and those have beem implemeted, such
as moving the labels closer the to the text boxs and making the margins
constistent with other tabs in the app. The closest weather tab was also
added to the page as that was orginally forgotten about from the photo
prototype.