Calculate the closest airport and get the weather
Implemented the calcultion for the closest airport. Big shout out to the
nearby airport group as we did a code review with them and they said it
was okay to modify their code for our needs which saved a lot of time.
After getting the closest airport, it now calls the api of the Aviation
Weather Center to get the metar and taf. Still need to implement getting
the coordinates from the actual device.