Got NearbyAirports working (again)
There was a bug that broke NearbyAirports. AirportToMilesConverter was
supposed to implement IValueConverter, which is part of .NET MAUI. But
the precursor to this port was console based, so I defined my own version
of IValueConverter so that it would compile. When that was deleted in this
app, the .NET MAUI version was used instead, and now NearbyAirports displays
mileage.

Also, there was a typo, Convertor => Converter