Map is now fully functional on both iOS and Androido
To summarize the last few commits, with .NET 8, the HeaderTemplate
would not appear, in iOS, in the Airports tab. I filed an issue,
and the person said that it worked in .NET 9. Upgrading to .NET 9
fixed that, but then it broke the Map page, as Mapsui 4.7 wouldn't
run on .NET 9. Upgrading to 5.0.0-beta.7, the map would show up
perfectly on RoutingStrategies, where they embeeded a Mapsui MapControl
in xaml. It did not show up at all on Maps, where they added a MapControl
programmatically as the first (and only) element of a grid.

It turned out that MapControl.UseGPU=false helped, sorta, kinda, some of
the time (at least periodically the map would show up, but it wasn't a sure
thing). But, putting the Mapsui MapControl directly into the xaml worked
(and was less complex).

Another issue arose with the RadioButtons -- there's an issue setting IsChecked
to True both in xaml and programatically, but the map shows up, rock solid (and
when Airports are added/deleted, the Map immediately reflects that.

	modified:   UI/MapPage.xaml
	modified:   UI/MapPage.xaml.cs