Added a header to the suggested route list of the routing strategies page.
Made the headers for the nearby airports and routing strategies pages sticky; they also are not visible until the user submits data for those pages.

The airport descriptions in the suggested route on the routing strategies page no longer overlap.

Keyboard is now hidden when the user submits data on the nearby airports and routing strategies pages.