Fixed CalculateNearbyAirports() to work on iOS
Addressed an interesting question of timing, code worked
correctly on Android but not iOS. We had to guarantee that
ConvertAll() was called before any airports were added, because
once any airport was added, iOS decided to trigger a redraw of the
CollectionView. The original code called ConvertAll() after all the
nearby airports had been added, and that was too late.

Fixed a typo in the CollectionView, an errant 1 was being printed
in the distance column

Also deleted numerous Console.WriteLine() debugging statements

	modified:   Model/BusinessLogic.cs
	modified:   Model/Database.cs
	modified:   Model/NearbyAirportBusinessLogic.cs
	modified:   Model/NearbyAirports/AirportToMilesConverter.cs
	modified:   UI/EnterAirportDetailsPopup.xaml.cs
	modified:   UI/MainTabbedPage.xaml.cs
	modified:   UI/MapPageSimple.xaml.cs
	modified:   UI/NearbyAirportsPage.xaml